Move waitForChannel state to find-or-create-ledger protocol
