+++ packages/wallet-protocols/src/protocols/ledger-funding/protocol.ts
@@ -1,4 +1,4 @@
-import { assign, DoneInvokeEvent, Machine, MachineConfig } from 'xstate';
+import { DoneInvokeEvent, Machine, MachineConfig } from 'xstate';
 import { Allocation } from '@statechannels/nitro-protocol';
 
 import { allocateToTarget, getEthAllocation } from '../../calculations';
@@ -11,70 +11,15 @@ import { CreateNullChannel, DirectFunding, SupportState } from '..';
 const PROTOCOL = 'ledger-funding';
 
 export interface Init {
+  ledgerChannelId: string;
   targetChannelId: string;
   deductions: Allocation;
 }
 
-/*
-My wallet's rule is to have at most one ledger channel open with any given peer.
-Therefore, two correct wallets should agree on which existing ledger channel, if any, to use
-in order to fund the target channel.
-
-A peer is identified by their participantId.
-*/
-
-const assignLedgerChannelId = assign(
-  (ctx: Init, event: DoneInvokeEvent<{ channelId: string }>) => ({
-    ...ctx,
-    ledgerChannelId: event.data.channelId,
-  })
-);
-
-const lookForExistingChannel = {
-  invoke: {
-    src: 'findLedgerChannelId',
-    onDone: [
-      {
-        target: 'success',
-        cond: 'channelFound',
-        actions: assignLedgerChannelId,
-      },
-      { target: 'determineLedgerChannel' },
-    ],
-  },
-};
-
-const determineLedgerChannel = {
-  invoke: {
-    src: 'getNullChannelArgs',
-    onDone: 'createNewLedger',
-  },
-};
-
-const createNewLedger = {
-  invoke: {
-    src: 'createNullChannel',
-    data: (_, { data }: DoneInvokeEvent<CreateNullChannel.Init>) => data,
-    onDone: { target: 'success', actions: assignLedgerChannelId },
-  },
-};
-
-const waitForChannel = {
-  initial: 'lookForExistingChannel',
-  states: {
-    lookForExistingChannel,
-    determineLedgerChannel,
-    createNewLedger,
-    success,
-  },
-  onDone: { target: 'fundLedger' },
-};
-
-type LedgerExists = Init & { ledgerChannelId: string };
 const fundLedger = {
   invoke: {
     src: 'directFunding',
-    data: ({ ledgerChannelId, deductions }: LedgerExists): DirectFunding.Init => ({
+    data: ({ ledgerChannelId, deductions }: Init): DirectFunding.Init => ({
       channelId: ledgerChannelId,
       minimalAllocation: deductions,
     }),
@@ -93,7 +38,6 @@ export const config: MachineConfig<any, any, any> = {
   key: PROTOCOL,
   initial: 'waitForChannel',
   states: {
-    waitForChannel,
     fundLedger,
     fundTarget,
     updateFunding,
@@ -107,8 +51,8 @@ export type Services = {
   getNullChannelArgs(ctx: Init): Promise<CreateNullChannel.Init>;
   createNullChannel: any;
   directFunding: any;
-  getTargetOutcome(ctx: LedgerExists): Promise<SupportState.Init>;
-  updateFunding(ctx: LedgerExists): Promise<void>;
+  getTargetOutcome(ctx: Init): Promise<SupportState.Init>;
+  updateFunding(ctx: Init): Promise<void>;
   supportState: ReturnType<typeof SupportState.machine>;
 };
 
@@ -134,7 +78,7 @@ export const machine: MachineFactory<Init, any> = (store: Store, context: Init) 
     targetChannelId,
     ledgerChannelId,
     deductions,
-  }: LedgerExists): Promise<SupportState.Init> {
+  }: Init): Promise<SupportState.Init> {
     const { latestState: ledgerState } = store.getEntry(ledgerChannelId);
 
     const ledgerAllocation = getEthAllocation(ledgerState.outcome, store.ethAssetHolderAddress);
@@ -153,7 +97,7 @@ export const machine: MachineFactory<Init, any> = (store: Store, context: Init) 
     };
   }
 
-  async function updateFunding({ targetChannelId, ledgerChannelId }: LedgerExists) {
+  async function updateFunding({ targetChannelId, ledgerChannelId }: Init) {
     const funding: Funding = { type: 'Indirect', ledgerId: ledgerChannelId };
     await store.setFunding(targetChannelId, funding);
   }
